Add feature
We can now interact to gain reputation and mney but loose energie. Energie is restore when we rest in the appatment. When max reputation is reached we lvl up and it required more reputation to lvl up again. Time pass when we work and sleep and a pop up message is display when trying to work with not enaught energie.